full refactoring
