+++ db_scripts/create_gs_db.sql
