+++ sf_telegram_bot.py
@@ -1,45 +1,8 @@
-import requests
 import PASSWORDS
 import sys
-from DBPostgres import DBPostgres
-from alert_to_mail import send_error_to_admin, get_participant_notification_text
-
-
-class TelegramBot:
-    """
-    Простейший класс Telegram бота
-    """
-    def __init__(self, bot_url, logger_, database=None):
-        self.bot_url = bot_url
-        self.logger = logger_
-        self.database = database
-
-    def get_text_updates(self, offset=0):
-        """
-        Get updates from Telegram
-        :return: Dictionary Updates
-        """
-        params = {'timeout': 1, 'offset': offset, 'limit': 1000, 'allowed_updates': ['message', 'edited_message']}
-        response = requests.get(self.bot_url + '/getUpdates', data=params)
-        self.logger.debug(response)
-        self.logger.debug(response.json())
-        if response.ok:
-            self.logger.debug("get_updates_json OK")
-            return True, response.json()
-        else:
-            self.logger.error(f"ERROR:{response.text}")
-            return False, response.text
-
-    def send_text_message(self, chat, text):
-        params = {'chat_id': chat, 'text': text}
-        response = requests.post(self.bot_url + '/sendMessage', data=params)
-        self.logger.debug(response)
-        if response.ok:
-            self.logger.debug("get_updates_json ОК")
-            return True, None
-        else:
-            self.logger.error(f"ERROR:{response.text}")
-            return False, response.text
+import Class_TelegramBot
+from Class_DBPostgres import DBPostgres
+from alert_to_mail import raise_error, get_participant_notification_text
 
 
 def mark_telegram_update_id(telegram_update_id_, logger_):
@@ -56,17 +19,6 @@ def mark_telegram_update_id(telegram_update_id_, logger_):
         raise_error("Не могу обновить telegram_update_id", logger_)
 
 
-def raise_error(err_text_, logger_):
-    """
-    Записать в лог ошибку - Выслать ошибку админу - Генерировать исключение
-    :param err_text_: Текст ошибки
-    :param logger_: логгер
-    """
-    logger_.error(err_text_)
-    send_error_to_admin(err_text_, logger_, prog_name="sf_telegram_bot.py")
-    raise Exception(err_text_)
-
-
 if __name__ == '__main__':
     import custom_logger
     import os
@@ -74,10 +26,10 @@ if __name__ == '__main__':
     logger = custom_logger.get_logger(program_file=os.path.realpath(__file__))
     logger.info("Try connect to DB")
     try:
-        dbconnect = DBPostgres(dbname=PASSWORDS.logins['postgres_dbname'], user=PASSWORDS.logins['postgres_user'],
-                               password=PASSWORDS.logins['postgres_password'],
-                               host=PASSWORDS.logins['postgres_host'],
-                               port=PASSWORDS.logins['postgres_port'], logger=logger)
+        dbconnect = DBPostgres(dbname=PASSWORDS.settings['postgres_dbname'], user=PASSWORDS.settings['postgres_user'],
+                               password=PASSWORDS.settings['postgres_password'],
+                               host=PASSWORDS.settings['postgres_host'],
+                               port=PASSWORDS.settings['postgres_port'], logger=logger)
     except Exception:
         raise_error("Can't connect to DB!!!", logger)
         sys.exit(1)
@@ -92,7 +44,7 @@ if __name__ == '__main__':
     except:
         raise_error("Не могу получить telegram_update_id из БД", logger)
     logger.debug(f"telegram_update_id={telegram_update_id}")
-    tb = TelegramBot(PASSWORDS.logins['telegram_bot_url'], logger)
+    tb = Class_TelegramBot.TelegramBot(PASSWORDS.settings['telegram_bot_url2'], logger)
     logger.info("Get updates")
     success, updates = tb.get_text_updates(telegram_update_id)
     logger.debug(f"success={success}")
